+++ extensions/standard-processors/processors/TextFragmentUtils.h
